feat: remove estate and parcel details (unused)
