+++ webapp/src/components/EstateDetailPage/EstateDetailPage.container.js
@@ -1,28 +1,22 @@
-import { withRouter } from 'react-router'
 import { connect } from 'react-redux'
+import { navigateTo } from '@dapps/modules/location/actions'
 
 import { locations } from 'locations'
-import { getMatchParams } from 'modules/location/selectors'
-import { getData as getParcels } from 'modules/parcels/selectors'
+import { getData as getAtlas } from 'modules/map/selectors'
 import { getData as getPublications } from 'modules/publication/selectors'
-import { areParcelsLoaded } from 'modules/estates/selectors'
-import { navigateTo } from '@dapps/modules/location/actions'
 
 import EstateDetailPage from './EstateDetailPage'
 
 const mapState = (state, ownProps) => {
-  const { id } = getMatchParams(ownProps)
-
   return {
-    id,
+    estate: ownProps.asset,
     publications: getPublications(state),
-    allParcels: getParcels(state),
-    isLoadingEstateParcels: !areParcelsLoaded(state, { id })
+    atlas: getAtlas(state)
   }
 }
 
 const mapDispatch = (dispatch, ownProps) => {
-  const { id } = getMatchParams(ownProps)
+  const id = ownProps.asset.id
 
   return {
     onEditParcels: () => dispatch(navigateTo(locations.editEstateParcels(id))),
@@ -34,4 +28,4 @@ const mapDispatch = (dispatch, ownProps) => {
   }
 }
 
-export default withRouter(connect(mapState, mapDispatch)(EstateDetailPage))
+export default connect(mapState, mapDispatch)(EstateDetailPage)
