+++ webapp/src/components/ParcelDetailPage/ParcelDescription/ParcelDescription.css
