--- webapp/src/components/ParcelDetailPage/ParcelDetail/index.js
@@ -1,3 +1,0 @@
-import ParcelDetail from './ParcelDetail'
-
-export default ParcelDetail
