+++ webapp/src/components/ParcelDetailPage/ParcelDetailPage.css
@@ -2,12 +2,12 @@
   padding-top: 0;
 }
 
-.ParcelDetailPage .ParcelDetail {
+.ParcelDetailPage {
   padding-top: 35px;
   font-size: 15px;
 }
 
-.ParcelDetail .ParcelName .name {
+.ParcelDetailPage .ParcelName .name {
   margin-right: 0.5em;
 }
 
@@ -15,26 +15,26 @@
   text-align: right !important;
 }
 
-.ParcelDetail .parcel-detail-row {
+.ParcelDetailPage .parcel-detail-row {
   margin-bottom: 2em;
 }
 
-.ParcelDetail .parcel-actions-container {
+.ParcelDetailPage .parcel-actions-container {
   display: flex !important;
   flex-flow: row nowrap;
   justify-content: flex-end;
   align-items: center;
 }
 
-.ParcelDetail h3 {
+.ParcelDetailPage h3 {
   font-family: var(--font-family) !important;
 }
 
-.ParcelDetail .mana-price-icon {
+.ParcelDetailPage .mana-price-icon {
   font-size: 1.5em;
 }
 
-.ParcelDetail .time-left .PublicationExpiration {
+.ParcelDetailPage .time-left .PublicationExpiration {
   font-size: 20px;
   display: block;
   color: var(--gray);
