+++ webapp/src/modules/estates/sagas.js
@@ -1,6 +1,6 @@
 import {
   takeEvery,
-  takeLatest,
+  // takeLatest,
   put,
   select,
   call,
@@ -19,7 +19,7 @@ import {
   REMOVE_PARCELS,
   DELETE_ESTATE_REQUEST,
   TRANSFER_ESTATE_REQUEST,
-  FETCH_ESTATE_SUCCESS,
+  // FETCH_ESTATE_SUCCESS,
   createEstateSuccess,
   createEstateFailure,
   fetchEstateSuccess,
@@ -33,16 +33,16 @@ import {
   transferEstateSuccess,
   transferEstateFailure
 } from 'modules/estates/actions'
-import { getEstates, areParcelsLoaded } from 'modules/estates/selectors'
-import { getData as getParcels } from 'modules/parcels/selectors'
+import { getEstates /*, areParcelsLoaded*/ } from 'modules/estates/selectors'
+// import { getData as getParcels } from 'modules/parcels/selectors'
 import { getAddress } from 'modules/wallet/selectors'
 import { api } from 'lib/api'
 import { encodeMetadata } from 'shared/asset'
 import { getParcelsNotIncluded } from 'shared/parcel'
 import { splitCoodinatePairs } from 'shared/coordinates'
 import { Bounds } from 'shared/map'
-import { fetchParcelRequest } from 'modules/parcels/actions'
-import { buildCoordinate } from 'shared/coordinates'
+// import { fetchParcelRequest } from 'modules/parcels/actions'
+// import { buildCoordinate } from 'shared/coordinates'
 
 export function* estateSaga() {
   yield takeEvery(CREATE_ESTATE_REQUEST, handleCreateEstateRequest)
@@ -51,7 +51,7 @@ export function* estateSaga() {
   yield takeEvery(EDIT_ESTATE_METADATA_REQUEST, handleEditEstateMetadataRequest)
   yield takeEvery(DELETE_ESTATE_REQUEST, handleDeleteEstate)
   yield takeEvery(TRANSFER_ESTATE_REQUEST, handleTransferRequest)
-  yield takeLatest(FETCH_ESTATE_SUCCESS, handleFetchEstateSuccess)
+  // yield takeLatest(FETCH_ESTATE_SUCCESS, handleFetchEstateSuccess)
 }
 
 function* handleCreateEstateRequest(action) {
@@ -216,15 +216,15 @@ function* handleTransferRequest({ estate, to }) {
   }
 }
 
-function* handleFetchEstateSuccess({ estate }) {
-  const areLoaded = yield select(state => areParcelsLoaded(state, estate))
-  if (!areLoaded) {
-    const parcels = yield select(state => getParcels(state))
-    for (const { x, y } of estate.data.parcels) {
-      const isLoaded = buildCoordinate(x, y) in parcels
-      if (!isLoaded) {
-        yield put(fetchParcelRequest(x, y))
-      }
-    }
-  }
-}
+// function* handleFetchEstateSuccess({ estate }) {
+//   const areLoaded = yield select(state => areParcelsLoaded(state, estate))
+//   if (!areLoaded) {
+//     const parcels = yield select(state => getParcels(state))
+//     for (const { x, y } of estate.data.parcels) {
+//       const isLoaded = buildCoordinate(x, y) in parcels
+//       if (!isLoaded) {
+//         yield put(fetchParcelRequest(x, y))
+//       }
+//     }
+//   }
+// }
